file renderes was removed and method object assign was changed
